+++ src/main/java/org/folio/entlinks/controller/delegate/AuthorityServiceDelegate.java
@@ -4,16 +4,21 @@ import static org.folio.entlinks.service.consortium.propagation.ConsortiumPropag
 import static org.folio.entlinks.service.consortium.propagation.ConsortiumPropagationService.PropagationType.DELETE;
 import static org.folio.entlinks.service.consortium.propagation.ConsortiumPropagationService.PropagationType.UPDATE;
 
+import java.util.List;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.folio.entlinks.controller.converter.AuthorityMapper;
 import org.folio.entlinks.domain.dto.AuthorityDto;
 import org.folio.entlinks.domain.dto.AuthorityDtoCollection;
+import org.folio.entlinks.domain.entity.Authority;
+import org.folio.entlinks.exception.RequestBodyValidationException;
+import org.folio.entlinks.integration.dto.event.DomainEventType;
 import org.folio.entlinks.service.authority.AuthorityDomainEventPublisher;
 import org.folio.entlinks.service.authority.AuthorityService;
 import org.folio.entlinks.service.consortium.propagation.ConsortiumAuthorityPropagationService;
 import org.folio.spring.FolioExecutionContext;
+import org.folio.tenant.domain.dto.Parameter;
 import org.springframework.stereotype.Service;
 
 @Log4j2
@@ -48,7 +53,10 @@ public class AuthorityServiceDelegate {
 
   public void updateAuthority(UUID id, AuthorityDto authorityDto) {
     var modifiedEntity = mapper.toEntity(authorityDto);
-    var oldDto = this.getAuthorityById(id);
+    var oldEntity = service.getById(id);
+    validateModifyPossibility(DomainEventType.UPDATE, oldEntity);
+
+    var oldDto = mapper.toDto(oldEntity);
     var updatedEntity = service.update(id, modifiedEntity);
     var newDto = mapper.toDto(updatedEntity);
     eventPublisher.publishUpdateEvent(oldDto, newDto);
@@ -57,9 +65,18 @@ public class AuthorityServiceDelegate {
 
   public void deleteAuthorityById(UUID id) {
     var entity = service.getById(id);
+    validateModifyPossibility(DomainEventType.DELETE, entity);
+
     var dto = mapper.toDto(entity);
     service.deleteById(id);
     eventPublisher.publishSoftDeleteEvent(dto);
     propagationService.propagate(entity, DELETE, context.getTenantId());
+  }
+
+  private void validateModifyPossibility(DomainEventType eventType, Authority entity) {
+    if (entity.isConsortiumShadowCopy()) {
+      throw new RequestBodyValidationException(eventType.name() + " is not applicable to consortium shadow copy",
+          List.of(new Parameter("id").value(String.valueOf(entity.getId()))));
+    }
   }
 }